AA-219: Added automatic versioning (version + build number + time) to index file
These changes enables grunt to create a new version.json file every time "grunt serve"
or "grunt build" is run. The new file contains a json with version (from bower.json)
the short hash of the tip of the current branch against which the grunt command has
been run, the date and time it was run.

The collected information is used to create a string that is displayed as version at
the bottom of the application.